+++ domain/models/room.go
@@ -7,8 +7,8 @@ import (
 // Room : roomテーブルモデル
 type Room struct {
 	ID        int64
-	RoomOwner int
-	GameTitle int
+	RoomOwner int64
+	GameTitle int64
 	Capacity  int
 	IsLock    bool
 	CreatedAt time.Time
